\name{StereoMorph-package}
\alias{StereoMorph-package}
\alias{StereoMorph}

\alias{dilateImage}
\alias{drawRectangle}
\alias{equalizeImageHist}
\alias{erodeImage}
\alias{findBoundaryPoints}
\alias{findCornerSubPix}
\alias{findEpipolarTangencyAngles}
\alias{generateQuads}
\alias{gridPointsFitError}
\alias{intCornersFromQuads}
\alias{meanBlurImage}
\alias{orderCorners}
\alias{rgbToGray}
\alias{thresholdImageMatrix}
\alias{inverseGridTransform}
\alias{dltCCEstimateStartParams}
\alias{listToJSONStr}
\alias{readXML4R}
\alias{readXMLLines}

\docType{package}
\title{Stereo Camera Morphometrics}
\description{
\strong{StereoMorph} provides functions for the collection of 3D points and curves using a stereo camera setup. \strong{StereoMorph} can also be used for \href{http://home.uchicago.edu/~aolsen/software/digitizing.shtml}{collecting 2D shape data from photographs}.
Please see \href{http://stereomorph.blogspot.com/p/stereomorph-10-step-tutorial.html}{StereoMorph 10-step Tutorial} for a step-by-step tutorial on how to use \strong{StereoMorph}.
}
\details{
\tabular{ll}{
Package: \tab StereoMorph\cr
Type: \tab Package\cr
Version: \tab 1.4\cr
Date: \tab 2015-04-06\cr
License: \tab GPL-2\cr
}
}
\author{
Aaron Olsen, Annat Haber

Maintainer: Aaron Olsen \email{aolsen@uchicago.edu}
}
\keyword{ package }
\seealso{
\href{http://cran.r-project.org/web/packages/svgViewR/index.html}{svgViewR}, 
\href{http://cran.r-project.org/web/packages/bezier/index.html}{bezier}
}
