\name{StereoMorph-package}
\alias{StereoMorph-package}
\alias{StereoMorph}

\alias{dilateImage}
\alias{drawRectangle}
\alias{equalizeImageHist}
\alias{erodeImage}
\alias{findBoundaryPoints}
\alias{findCornerSubPix}
\alias{findEpipolarTangencyAngles}
\alias{generateQuads}
\alias{gridPointsFitError}
\alias{intCornersFromQuads}
\alias{meanBlurImage}
\alias{orderCorners}
\alias{rgbToGray}
\alias{thresholdImageMatrix}

\docType{package}
\title{Stereo Camera Morphometrics}
\description{
\strong{StereoMorph} provides functions for the collection of 3D points and curves using a stereo camera setup. \strong{StereoMorph} can also be used for collecting 2D shape data from photographs. Please see \url{http://home.uchicago.edu/~aolsen/StereoMorph/StereoMorph Tutorial.pdf} for a step-by-step tutorial on how to use \strong{StereoMorph}. Tutorials and other examples can also be found at \url{http://stereomorph.blogspot.com/}.
}
\details{
\tabular{ll}{
Package: \tab StereoMorph\cr
Type: \tab Package\cr
Version: \tab 1.2\cr
Date: \tab 2014-08-23\cr
License: \tab GPL-2\cr
}
}
\author{
Aaron Olsen, Annat Haber

Maintainer: Aaron Olsen \email{aolsen@uchicago.edu}
}
\keyword{ package }
\seealso{
\href{http://cran.r-project.org/web/packages/bezier/index.html}{bezier}
}
