\name{StereoMorph-package}
\alias{StereoMorph-package}
\alias{StereoMorph}
\docType{package}
\title{Stereo Camera Morphometrics}
\description{
\strong{StereoMorph} provides functions for the collection of 3D points and curves using a stereo camera setup.

A step-by-step tutorial on how to use \strong{StereoMorph} can be found at \url{http://stereomorph.blogspot.com/p/stereomorph-10-step-tutorial.html}.
}
\details{
\tabular{ll}{
Package: \tab StereoMorph\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-03-05\cr
License: \tab GPL-2\cr
}
}
\author{
Aaron Olsen, Annat Haber

Maintainer: Aaron Olsen \email{aolsen@uchicago.edu}
}
\keyword{ package }
\seealso{
\href{http://cran.r-project.org/web/packages/bezier/index.html}{bezier}
}
