\name{StereoMorph-package}
\alias{StereoMorph-package}
\alias{StereoMorph}

\alias{dilateImage}
\alias{drawRectangle}
\alias{equalizeImageHist}
\alias{erodeImage}
\alias{findBoundaryPoints}
\alias{findCornerSubPix}
\alias{findEpipolarTangencyAngles}
\alias{generateQuads}
\alias{gridPointsFitError}
\alias{intCornersFromQuads}
\alias{meanBlurImage}
\alias{orderCorners}
\alias{rgbToGray}
\alias{thresholdImageMatrix}
\alias{inverseGridTransform}
\alias{dltCCEstimateStartParams}

\docType{package}
\title{Stereo Camera Morphometrics}
\description{
\strong{StereoMorph} provides functions for the collection of 3D points and curves using a stereo camera setup. \strong{StereoMorph} can also be used for collecting 2D shape data from photographs. Please see \url{http://stereomorph.blogspot.com/p/stereomorph-10-step-tutorial.html} for a step-by-step tutorial on how to use \strong{StereoMorph}.
}
\details{
\tabular{ll}{
Package: \tab StereoMorph\cr
Type: \tab Package\cr
Version: \tab 1.3\cr
Date: \tab 2015-02-11\cr
License: \tab GPL-2\cr
}
}
\author{
Aaron Olsen, Annat Haber

Maintainer: Aaron Olsen \email{aolsen@uchicago.edu}
}
\keyword{ package }
\seealso{
\href{http://cran.r-project.org/web/packages/bezier/index.html}{bezier}
}
